Split config module in multiple files
In preparation of the additional configuration items required by the new
rate control api and gop-parallel API.
